fix(rng): 补充审计二轮六项——重放零副作用/NextInt 全跨度/API clamp/条目上限/同流异seed失败关闭（#1039）
- VerifyReplay 前后保存/恢复活流位置：确定性证明不再回卷共享流（消除后续 auto burst 与录制段重复输出的观感）
- RngStream.NextInt 改 long 基算术：全 int 跨度 [MinValue, MaxValue) 不再回绕越下界
- GasGraphRuntimeApi.WeightedPick 在 API 面 clamp modulation permille 到 [-1000,1000]（与图 op 注释合同对齐）
- DistributionTable 条目数上限 1024（stackalloc 有界，超限失败关闭）
- DistributionConfigLoader：同流异 streamSeed 由静默首胜改为抛可读错误（无隐式回退）
- 遗留未做（记档）：enabled 缺省 false 的直觉问题（DTO 默认值受构造顺序约束，需连测试改签名，随 #1045 切片处理）

+++ src/Core/Engine/Randomization/RngStream.cs
@@ -58,7 +58,7 @@ namespace Ludots.Core.Engine.Randomization
                 throw new ArgumentOutOfRangeException(nameof(maxExclusive), "Upper bound must be greater than the lower bound.");
             }
 
-            var range = (uint)(maxExclusive - minInclusive);
+            var range = (uint)((long)maxExclusive - minInclusive);
             var limit = uint.MaxValue - uint.MaxValue % range;
             uint sample;
             do
@@ -66,7 +66,7 @@ namespace Ludots.Core.Engine.Randomization
                 sample = NextUInt();
             } while (sample >= limit);
 
-            return minInclusive + (int)(sample % range);
+            return (int)(minInclusive + (long)(sample % range));
         }
 
         public void Advance(int steps)
